refactor: 抽离 debounce 函数到公共函数中
