+++ lib/st_video/video_fullscreen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 
 import 'package:saturn/st_icons/st_icons.dart';
 import 'package:saturn/st_video/video_control.dart';
-import 'package:saturn/st_video/video_debounce.dart';
+import 'package:saturn/utils/debounce.dart';
 import 'package:saturn/st_video/video_progress.dart';
 import 'package:saturn/st_video/video_sound.dart';
 import 'package:saturn/st_video/video_util.dart';
